[Setup] Ignore 'sodium' on client side bundling.
The transient `keybase-nacl` dependency includes two version of
libsodium: `sodium` for server side testing and `tweetnacl` as
lightweight alternative for browser bundles.

For some reasons browserify tries to include the node-js only
`sodium` module which results in an 'module not found' error
message as long as we don't exclude it.